+++ libCacheSim/cache/eviction/LeCaR.c
@@ -533,10 +533,10 @@ void LeCaR_evict(cache_t *cache, const request_t *req,
   obj_to_evict->LeCaR.eviction_vtime = cache->n_req;
 
 #ifdef TRACK_EVICTION_R_AGE
-  record_eviction_age(cache, (int)(req->real_time - obj_to_evict->create_time));
+  record_eviction_age(cache, req->real_time - obj_to_evict->create_time);
 #endif
 #ifdef TRACK_EVICTION_V_AGE
-  record_eviction_age(cache, (int)(cache->n_req - obj_to_evict->create_time));
+  record_eviction_age(cache, cache->n_req - obj_to_evict->create_time);
 #endif
 
   if (evicted_obj != NULL) {
